refactor(gpui): split command_palette/mod.rs into role-based modules
mod.rs went from 243 lines holding everything to 8 lines of pub mod /
pub use. Each new file is focused:

- state.rs (27)  — CommandPalette struct, CommandOutput enum,
                   Focusable impl
- input.rs (140) — open(), parse_command(), matches(), on_key()
                   (the input/parse handlers)
- exec.rs (52)   — run_command() + execute() (process spawn + output
                   capture, isolated from rendering)
- view.rs (37)   — impl Render

actions.rs and render.rs (rendering helpers) keep their existing
shape. CommandPalette stays the single public type via mod.rs's
re-export, so external callers (CommandPalette::open) are unchanged.
