+++ shell/gpui/src/windows/command_palette/view.rs
@@ -1,0 +1,37 @@
+use gpui::{
+    Context, InteractiveElement, IntoElement, ParentElement, Render, Styled, Window, div, rgb,
+};
+
+use super::render::{action_list, command_view, divider, query_box};
+use super::state::{CommandOutput, CommandPalette};
+use crate::app::theme::theme;
+
+impl Render for CommandPalette {
+    fn render(&mut self, _window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
+        let t = theme(cx).clone();
+        let visible = self.matches();
+        let selected = self.selected.min(visible.len().saturating_sub(1));
+
+        let body = match (&self.output, self.parse_command()) {
+            (CommandOutput::Idle, None) => action_list(&visible, selected, &t).into_any_element(),
+            (CommandOutput::Idle, Some(body)) => {
+                let cmd = format!("jj {body}");
+                command_view(None, cmd.trim_end(), &t).into_any_element()
+            }
+            (out, _) => command_view(Some(out), "", &t).into_any_element(),
+        };
+
+        div()
+            .track_focus(&self.focus_handle)
+            .key_context("CommandPalette")
+            .on_key_down(cx.listener(Self::on_key))
+            .flex()
+            .flex_col()
+            .size_full()
+            .bg(rgb(t.detail_bg))
+            .text_color(rgb(t.fg))
+            .child(query_box(&self.query, &t))
+            .child(divider(&t))
+            .child(body)
+    }
+}
